Handle empty output and no explicit -out flag
The tool assumed we were using the `-out` flag to specify an output
file, which would result in `Path: ...` appearing in Terraform's output.

Since we don't want to require the use of the `-out` flag, improve the
preface detection to handle this case. Also handle the case where there
are no changes detected.