Update the support for unquoted bracketed output.
This updates the previous change to handle Terraform outputting `<computed>`
fields without quotes to include other possible values. This was also breaking
for `<sensitive>` fields, such as when launching new RDS instances and defining
the password.

This changes it to just handle generic content so it isn't hardcoded to specific
values.