Fix Ruby 3 syntax errors
The Ruby code generated by Treetop was creating methods with names
such as `_1` which are the tokens that correspond to the `_` rule.
This causes Ruby 2.7 to print warnings that `_1`, `_2`, etc. are
reserved for numbered parameters. In Ruby 3, they are syntax errors.

This commit renames `_` to `ws` so that the generated methods now
have names such as `ws1`, `ws2`, etc.